Added the 'results.split' option in order to break a large ResultSet into smaller ResultSet's

+++ src/main/java/nl/anchormen/sql4es/ESResultSet.java
@@ -1376,7 +1376,7 @@ public class ESResultSet implements ResultSet {
 		
 	}
 
-	public ESResultSet setOffset(int offset) {
+	public ESResultSet setOffset(long offset) {
 		this.offset = offset;
 		return this;
 	}
